Move class into its own file
